Restructure node editor e2e tests
